+++ packages/ui/input/checkbox/components/switch.module.css
