+++ stories/alerts/alerts.mdx
@@ -1,8 +1,9 @@
 import { useState, useEffect, useCallback, useMemo } from "react";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
 import { Alerts, pushAlert, clearAlerts } from "@digital-science/figshare-fcl/alerts";
 import { Button } from "@digital-science/figshare-fcl/button";
 import { Cover } from "../story-utils/Cover";
+
+import { Meta, ArgTypes } from "@storybook/blocks";
 
 <Meta
   title="UI/Alerts/Alerts"
