+++ stories/alerts/fieldError.mdx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback, useMemo } from "react";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import TextInput from "@digital-science/figshare-fcl/input/text";
 import { FieldError } from "@digital-science/figshare-fcl/fieldError";
 import { GenericButton } from "@digital-science/figshare-fcl/button";
@@ -87,7 +87,6 @@ Usually tied to inputs or other form fields in order to show various validation 
 
 ```jsx
 import React, { useState, useEffect, useCallback, useMemo } from "react";
-import { Meta,  Canvas, ArgTypes } from '@storybook/blocks';
 import TextInput from "@digital-science/figshare-fcl/input/text";
 import { FieldError } from "@digital-science/figshare-fcl/fieldError";
