+++ stories/avatar/avatar.mdx
@@ -1,10 +1,11 @@
 import { useState, useEffect, useCallback, useMemo } from "react";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import { Avatar } from "@digital-science/figshare-fcl/avatar";
 import { Button } from "@digital-science/figshare-fcl/button";
 import image from "../../assets/avatar.png";
 import imageWithTransparentBackground from "../../assets/avatar-transparent-background.png";
 import { Cover } from "../story-utils/Cover";
+
 
 export const Default = () => (
   <div style={{ display: "flex", flexDirection: "column", gap: "12px" }}>
