+++ stories/button/button/index.mdx
@@ -1,7 +1,9 @@
 import Home from "@digital-science/figshare-fcl/icons/home/medium";
-import Button from "@digital-science/figshare-fcl/button/button";
+import { Button } from "@digital-science/figshare-fcl/button";
 import { Cover } from "../../story-utils/Cover";
 
+
+import { Meta, ArgTypes } from "@storybook/blocks";
 
 <Meta 
   title="UI/Button/Button"
