+++ stories/button/genericButton/index.mdx
@@ -3,6 +3,8 @@ import GenericButton from "@digital-science/figshare-fcl/button/genericButton";
 import { Cover } from "../../story-utils/Cover";
 
 
+import { Meta, ArgTypes } from "@storybook/blocks";
+
 <Meta 
   title="UI/Button/Generic Button" 
   component={GenericButton}
