+++ stories/button/iconButton/index.mdx
@@ -3,6 +3,8 @@ import IconButton from "@digital-science/figshare-fcl/button/iconButton";
 import { Cover } from "../../story-utils/Cover";
 
 
+import { Meta, ArgTypes } from "@storybook/blocks";
+
 <Meta 
   title="UI/Button/Icon Button" 
   component={IconButton}
