+++ stories/fieldset/fieldset.mdx
@@ -1,10 +1,11 @@
 import { useState, useCallback } from 'react';
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import Radio from "@digital-science/figshare-fcl/input/radio";
 import { Button } from "@digital-science/figshare-fcl/button";
 import { Fieldset, Legend, Items } from "@digital-science/figshare-fcl/fieldset";
 import Input from "@digital-science/figshare-fcl/input";
 import { Cover } from "../story-utils/Cover";
+
 
 <Meta
   title="UI/Fieldset"
