+++ stories/globals/colors.mdx
@@ -1,8 +1,9 @@
 import classnames from "classnames";
 import { Fragment } from "react";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import styles from "./colors.stories.css";
 import { Cover } from "../story-utils/Cover";
+
 
 
 export const onSaveTitleClipboard = (ev) => {
