+++ stories/globals/typography.mdx
@@ -1,7 +1,8 @@
 import classnames from "classnames";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import styles from "./typography.stories.css";
 import { Cover } from "../story-utils/Cover";
+
 
 <Meta 
   title="Global Styles/Typography"
