+++ stories/icons/figshare.mdx
@@ -4,6 +4,8 @@ import FigshareFullLogo from "@digital-science/figshare-fcl/icons/figshare/index
 import { Cover } from "../story-utils/Cover";
 
 
+import { Meta } from "@storybook/blocks";
+
 export const Container = (props) => {
   const style = {
     display: "flex",
