+++ stories/icons/icons.mdx
@@ -2,6 +2,8 @@
 import styles from "./icons.stories.css";
 import { Cover } from "../story-utils/Cover";
 
+import { Meta } from "@storybook/blocks";
+
 <Meta 
   title="Icons/Generic" 
   parameters={{
