+++ stories/icons/itemType/index.mdx
@@ -2,6 +2,8 @@
 import styles from "./index.stories.css";
 import { Cover } from "../../story-utils/Cover";
 
+import { Meta } from "@storybook/blocks";
+
 <Meta 
   title="Icons/ItemType" 
   parameters={{
