+++ stories/input/checkbox/checkbox.mdx
@@ -1,9 +1,10 @@
 import { useState, useCallback } from 'react';
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import TextInput from "@digital-science/figshare-fcl/input/text";
 import Checkbox from "@digital-science/figshare-fcl/input/checkbox";
 import Input from "@digital-science/figshare-fcl/input";
 import { Cover } from "../../story-utils/Cover";
+
 
 <Meta 
   title="UI/Inputs/Checkbox"
