+++ stories/input/lexical/index.mdx
@@ -4,6 +4,8 @@ import { Button } from "@digital-science/figshare-fcl/button";
 import Editor from "@digital-science/figshare-fcl/textEditor/Lexical";
 import { Cover } from "../../story-utils/Cover";
 
+import { Meta, ArgTypes } from "@storybook/blocks";
+
 export const canvasStyle = {
   display: "flex",
   flexDirection: "column",
