+++ stories/input/password/index.mdx
@@ -1,6 +1,8 @@
 import PasswordInput from "@digital-science/figshare-fcl/input/password";
 import { Cover } from "../../story-utils/Cover";
 
+import { Meta, ArgTypes } from "@storybook/blocks";
+
 <Meta 
   title="UI/Inputs/Password" 
   component={PasswordInput}
