+++ stories/input/radio/radio.mdx
@@ -1,8 +1,9 @@
 import { useState, useCallback } from 'react';
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import Radio from "@digital-science/figshare-fcl/input/radio";
 import Input from "@digital-science/figshare-fcl/input";
 import { Cover } from "../../story-utils/Cover";
+
 
 <Meta
   title="UI/Inputs/Radio"
