+++ stories/input/search/index.mdx
@@ -1,8 +1,9 @@
 
 import { useState, useCallback } from 'react';
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import SearchInput from "@digital-science/figshare-fcl/input/search";
 import { Cover } from "../../story-utils/Cover";
+
 
 <Meta 
   title="UI/Inputs/Search"
