+++ stories/input/text/index.mdx
@@ -2,6 +2,8 @@ import TextInput from "@digital-science/figshare-fcl/input/text";
 import Input from "@digital-science/figshare-fcl/input";
 import { Cover } from "../../story-utils/Cover";
 
+import { Meta, ArgTypes } from "@storybook/blocks";
+
 <Meta 
   title="UI/Inputs/Text" 
   component={Input}
