+++ stories/input/textarea/index.mdx
@@ -1,7 +1,8 @@
 import { useState, useCallback } from "react";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import Textarea from "@digital-science/figshare-fcl/textarea";
 import { Cover } from "../../story-utils/Cover";
+
 
 export const ControlledTextarea = (props) => {
   const [value, setValue] = useState("");
