+++ stories/navigableList/navigableList.mdx
@@ -1,6 +1,8 @@
 import NavigableList from "@digital-science/figshare-fcl/helpers/navigableList";
 import { Cover } from "../story-utils/Cover";
 
+import { Meta, ArgTypes } from "@storybook/blocks";
+
 export const containerStyle = {
   margin: "calc(3 * var(--gridSize))",
   padding: "calc(3 * var(--gridSize))",
