+++ stories/select/select.mdx
@@ -1,9 +1,10 @@
 import { useState } from "react";
 import classnames from "classnames";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import { Select, Option } from "../../packages/ui/select";
 import { Button } from "@digital-science/figshare-fcl/button";
 import { Cover } from "../story-utils/Cover";
+
 
 <Meta
  title="UI/Select"
