+++ stories/tabs/index.mdx
@@ -1,11 +1,12 @@
 import { useState, useEffect, useCallback, useMemo } from "react";
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import { Tabs, TabsList, Tab, TabsPanel, TabEntry } from "@digital-science/figshare-fcl/tabs";
 import { Button } from "@digital-science/figshare-fcl/button";
 import Edit from "@digital-science/figshare-fcl/icons/react/Edit";
 import image from "../../assets/avatar.png";
 import imageWithTransparentBackground from "../../assets/avatar-transparent-background.png";
 import { Cover } from "../story-utils/Cover";
+
 export const tab = "map";
 export const onTab = () => {};
