+++ stories/toggletip/index.mdx
@@ -1,11 +1,12 @@
 import { useState, useCallback } from 'react';
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import Info from "@digital-science/figshare-fcl/icons/info/medium";
 import { Button, IconButton, GenericButton} from "@digital-science/figshare-fcl/button";
 
 import { Toggletip, Trigger, Content, UncontrolledToggletip } from "@digital-science/figshare-fcl/toggletip";
 import { Cover } from "../story-utils/Cover";
 import styles from "./index.stories.css";
+
 
 <Meta
   component={Toggletip}
