+++ stories/tooltip/index.mdx
@@ -1,11 +1,12 @@
 import { useState, useCallback } from 'react';
-import { Meta, ArgTypes } from '@storybook/addon-docs';
+import { Meta, ArgTypes } from '@storybook/blocks';
 import { Button, GenericButton } from "@digital-science/figshare-fcl/button";
 import Info from "@digital-science/figshare-fcl/icons/info/medium";
 
 import { Content, Trigger, UncontrolledTooltip as Tooltip } from "@digital-science/figshare-fcl/tooltip";
 import { Cover } from "../story-utils/Cover";
 import styles from "./index.stories.css";
+
 
 <Meta
   component={Tooltip}
